Revert "Add vendor extensions"
